Initial structure for fix.
