+++ .gitignore
@@ -310,3 +310,5 @@ Contact GitHub API Training Shop Blog About
 out
 
 *.localdev.*
+
+/**/ToTest.html
